handle api with portal-doctor and display

+++ PetCare-Server/PetCare-Infotype/src/main/java/com/group2/petcareinfotype/entity/Examination.java
@@ -27,7 +27,7 @@ public class Examination implements Serializable {
 
     private Integer date;
 
-    private Integer time;
+    private Double time;
 
     private String firstDescription;
